+++ biblioteca-4/tablas/db-usuarios/modificar-3.php
@@ -18,7 +18,7 @@ if (hayErrores("modificar-3")) {
     exit();
 }
 
-compruebaAvisosGenerales("usuarios", "modificar-3", "registrosExisten", "usuarios", $id);
+compruebaAvisosGenerales("usuarios", "modificar-3", "registrosExisten", $id);
 
 if (hayErrores("modificar-3")) {
     header("Location:modificar-1.php");
@@ -31,7 +31,7 @@ if (hayErrores("modificar-3")) {
 
 compruebaAvisosGenerales("usuarios", "modificar-3", "yaExisteRegistroConOtroId", "usuario", "password", "nivel", "id");
 
-compruebaAvisosGenerales("usuarios", "modificar-3", "incluyeUsuarioRoot", $id);
+compruebaAvisosGenerales("usuarios", "modificar-3", "incluyeUsuarioRoot", $usuario, $password, $nivel, $id);
 
 if (hayErrores("modificar-3")) {
     incluyeValoresOriginalesEnAvisos("usuarios", "modificar-3", "usuario", "password", "nivel", "id");